==> 03_Fst_Plink_Input.R <==

#Get fst between Europe and North America, where Europe and North American individuals are cases vs controls
./plink --bfile Nor_Can_GOOD_SNPs_maf001 --fst case-control --allow-extra-chr --chr-set 30 --allow-no-sex
